apps/api/pyproject.toml
[project]
name = "finetech-api"
version = "0.1.0"
description = "FastAPI backend for the Finetech unified financial intelligence platform."
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.112.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.0.0",
  "sqlalchemy[asyncio]>=2.0.0",
  "asyncpg>=0.29.0",
  "alembic>=1.13.0",
  "python-dotenv>=1.0.0",
  "httpx>=0.27.0",
  "redis>=5.0.0",
  "structlog>=24.0.0",
  "shap>=0.46.0",
  "lime>=0.2.0.1",
  "torch>=2.2.0",
  "numpy>=1.26.0",
  "google-generativeai>=0.8.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["finetech_api*"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

